Add Forecastable module & include in ForecastShowFacade & RoadTripsCreateFacade
