+++ app/modules/directionable.rb
@@ -1,15 +1,15 @@
 module Directionable
   private
   
-  def google_api
+  def google_directions_api
     parameters = { origin: @origin, destination: @destination }
-    @google_api ||= ApiService::Google.new(parameters)
+    @google_directions_api ||= ApiService::Google.new(parameters)
   end
   
   def api_directions_hash
-    @api_directions_hash ||= google_api.directions
+    @api_directions_hash ||= google_directions_api.directions
   end
-
+  
   def duration_in_seconds
     api_directions_hash[:routes].first[:legs].first[:duration][:value]
   end
